[Build] Fix debug build on Linux
Summary: incompatible compiler flag and int unsigned int comparison

Test Plan: tested on ubuntu

Reviewers: ilijar, michsien, piotr, nand

Reviewed By: nand

Differential Revision: https://ec2-52-30-7-90.eu-west-1.compute.amazonaws.com/D186

+++ master/ProCamSystem.cc
@@ -24,7 +24,7 @@ cv::Mat loadMat(
     throw EXCEPTION() << "Array expected in JSON.";
   }
 
-  if (width * height * D != static_cast<int>(arr.size())) {
+  if (width * height * D != arr.size()) {
     throw EXCEPTION() << "Invalid array size.";
   }
   cv::Mat mat = cv::Mat::zeros({ width, height }, CVType<T, D>::type);
